Working version

+++ src/main/java/com/github/mcreeper12731/engine/evaluators/Evaluator.java
@@ -1,7 +1,7 @@
 package com.github.mcreeper12731.engine.evaluators;
 
-import com.github.mcreeper12731.game.logic.Game;
+import com.github.mcreeper12731.game.Game;
 
 public interface Evaluator {
-    int evaluate(Game game);
+    double evaluate(Game game);
 }
